+++ module-common/src/main/java/hellojpa/domain/BaseEntity.java
@@ -1,4 +1,4 @@
-package hellojpa;
+package hellojpa.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
