--- module-movie/src/main/java/hellojpa/Main.java
@@ -1,7 +1,0 @@
-package hellojpa;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
