+++ module-reservation/src/main/java/hellojpa/aop/DistributedLock.java
@@ -1,4 +1,4 @@
-package hellojpa;
+package hellojpa.aop;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
