+++ module-reservation/src/main/java/hellojpa/aop/DistributedLockAop.java
@@ -1,4 +1,4 @@
-package hellojpa;
+package hellojpa.aop;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -20,7 +20,7 @@ public class DistributedLockAop {
     private final RedissonClient redissonClient;
     private final AopForTransaction aopForTransaction;
 
-    @Around("@annotation(hellojpa.DistributedLock)")
+    @Around("@annotation(hellojpa.aop.DistributedLock)")
     public Object lock(final ProceedingJoinPoint joinPoint, DistributedLock distributedLock) throws Throwable {
 
         String key = REDISSON_LOCK_PREFIX + distributedLock.key();
