+++ module-reservation/src/main/java/hellojpa/aop/DistributedLockKeyGenerator.java
@@ -1,4 +1,4 @@
-package hellojpa;
+package hellojpa.aop;
 
 import org.springframework.expression.ExpressionParser;
 import org.springframework.expression.spel.standard.SpelExpressionParser;
