--- module-reservation/src/test/java/hellojpa/TestApplication.java
@@ -1,10 +1,0 @@
-package hellojpa;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TestApplication {
-    public static void main(String[] args) {
-
-    }
-}
