--- module-reservation/src/test/resources/application.yml
@@ -1,10 +1,0 @@
-spring:
-  datasource:
-    url: jdbc:mysql://localhost:3307/redisdb
-    username: user
-    password: user1234
-    driver-class-name: com.mysql.cj.jdbc.Driver
-  jpa:
-    hibernate:
-      ddl-auto: update
-    show-sql: true
